requirements.txt
torch
torchvision
gradio
numpy
pillow
pytorch-msssim
tqdm
